+++ Server/src/main/java/com/mikpuk/vavaserver/UserDAO.java
@@ -6,6 +6,6 @@ public interface UserDAO {
 
     public void setDataSource(DataSource dataSource);
     public void createUser(String name,String password);
-    public User getUserById(int id);
+    public User getUserById(long id);
     public User getUserByData(String username,String password);
 }
